% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_by_municipality.R
\docType{data}
\name{election_by_municipality}
\alias{election_by_municipality}
\title{Election results by municipality}
\format{
A data frame with 87,048 rows and 9 columns
\describe{
  \item{year}{Election year}
  \item{type}{"General" or "Primary". (Currently all rows are "General")}
  \item{office}{"President", "Senate", or "Governor"}
  \item{GEOID}{US Census GEOID for the municipality}
  \item{county}{County of the municipality}
  \item{municipality}{Municipality name}
  \item{candidate}{Candidate name}
  \item{party}{Candidate party}
  \item{vote}{Number of votes}
}
}
\source{
\url{https://nj.gov/state/elections/election-information-results.shtml}
}
\usage{
election_by_municipality
}
\description{
Election results by municipality
}
\keyword{datasets}
