% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_by_county.R
\docType{data}
\name{election_by_county}
\alias{election_by_county}
\title{Election results by county}
\format{
A data frame with 2,835 rows and 8 columns
\describe{
  \item{year}{Election year}
  \item{type}{"General" or "Primary". (Currently all rows are "General")}
  \item{office}{"President", "Senate", or "Governor"}
  \item{GEOID}{US Census GEOID for the county}
  \item{county}{County name}
  \item{candidate}{Candidate name}
  \item{party}{Candidate party}
  \item{vote}{Number of votes}
}
}
\source{
\url{https://nj.gov/state/elections/election-information-results.shtml}
}
\usage{
election_by_county
}
\description{
Election results by county
}
\keyword{datasets}
